src/Back-end/requirements.txt
django >= 4.0.3
Pillow >= 9.0.1
mysqlclient >= 2.1.0
django-crispy-forms >= 1.14.0
djangorestframework == 3.13.1
djangorestframework-simplejwt == 5.2.0